LayoutTests/fast/xpath/detached-subtree-invalidate-iterator-expected.txt

Test that iterators are invalidated even if the original context is detached.

PASS result.invalidIteratorState is false
PASS result.iterateNext().tagName is 'span'
Modifying the document...
PASS result.invalidIteratorState is true
PASS result.iterateNext() threw exception InvalidStateError (DOM Exception 11): The object is in an invalid state..
PASS successfullyParsed is true

TEST COMPLETE
